refactor: remove legacy code and clean up Rust modules
- Remove placeholder storage module
- Consolidate aug7 → dom7#5 alias (identical intervals)
- Remove redundant aug7 ChordQuality and chord_name arm
- Simplify Note::transpose (remove double cast)
- Fix test indentation in fretboard.rs
- Remove legacy Voicing struct and generate() function
- Extract fret_stats helper, replace inline span calculation
- Add VoicingRecipe::short_label(), make guide_tones pub
- Refactor shell/rootless generation to use shared guide_tones

+++ src/theory/notes.rs
@@ -24,10 +24,7 @@ impl Note {
 
     /// Transpose this note by a given number of semitones.
     pub fn transpose(self, semitones: i32) -> Self {
-        Self::new(
-            self.pitch_class as i32 + semitones,
-            self.octave as i32 as i8,
-        )
+        Self::new(self.pitch_class as i32 + semitones, self.octave)
     }
 
     /// The MIDI note number for this note.
